word/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httptools==0.6.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.6.0
odmantic==1.0.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.2
setuptools==75.1.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.40.0
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
wheel==0.44.0
